feat(electron): Move common ANR code to utils
